+++ eventmetrics/parsers/eventConfig.go
@@ -20,7 +20,6 @@ package parsers
 import (
 	"errors"
 	"regexp"
-	"strings"
 	"time"
 
 	"github.com/xmidt-org/interpreter"
@@ -88,70 +87,4 @@ func (e EventInfo) Valid(event interpreter.Event) (bool, error) {
 		return true, nil
 	}
 	return e.Validator.Valid(event)
-}
-
-// TimeLocation is an enum to determine which timestamp should be used in timeElapsed calculations
-type TimeLocation int
-
-const (
-	Birthdate TimeLocation = iota
-	Boottime
-)
-
-var (
-	timeLocationUnmarshal = map[string]TimeLocation{
-		"birthdate": Birthdate,
-		"boot-time": Boottime,
-	}
-)
-
-// ParseTimeLocation returns the TimeLocation enum when given a string.
-func ParseTimeLocation(location string) TimeLocation {
-	location = strings.ToLower(location)
-	if value, ok := timeLocationUnmarshal[location]; ok {
-		return value
-	}
-	return Birthdate
-}
-
-// ParseTime gets the time from the proper location of an Event
-func ParseTime(e interpreter.Event, location TimeLocation) time.Time {
-	if location == Birthdate {
-		if e.Birthdate > 0 {
-			return time.Unix(0, e.Birthdate)
-		} else {
-			return time.Time{}
-		}
-
-	}
-
-	if bootTime, err := e.BootTime(); err == nil {
-		return time.Unix(bootTime, 0)
-	} else {
-		return time.Time{}
-	}
-}
-
-// SessionType is an enum to determine which session an event should be searched from
-type SessionType int
-
-const (
-	Previous SessionType = iota
-	Current
-)
-
-var (
-	sessionTypeUnmarshal = map[string]SessionType{
-		"previous": Previous,
-		"current":  Current,
-	}
-)
-
-// ParseSessionType returns the SessionType enum when given a string.
-func ParseSessionType(session string) SessionType {
-	session = strings.ToLower(session)
-	if value, ok := sessionTypeUnmarshal[session]; ok {
-		return value
-	}
-	return Previous
 }
